refactor(sql-editor): sql_editor_dialog.py(3997줄)의 부속 클래스 6개 분리
TF-STATUS-006 대형 파일 watch 항목 후속 (Phase B). 핵심 SQLEditorDialog
클래스(~2480줄)는 쿼리실행/트랜잭션/결과셀편집이 상태를 깊이 공유하고 있어
이번 패스에서는 그대로 두고, 안전하게 분리 가능한 부속 클래스 6개만 추출.

- sql_editor_highlighters.py (신규): SQLHighlighter, SQLValidatorHighlighter
- sql_editor_autocomplete.py (신규): AutoCompletePopup
- sql_editor_code_editor.py (신규): LineNumberArea, CodeEditor,
  ValidatingCodeEditor, SQLEditorTab, LARGE_SQL_RENDER_LIMIT_BYTES
- sql_editor_history_dialog.py (신규): HistoryDialog
- sql_editor_workers.py (신규): SQLQueryWorker, SQLTransactionExecutionWorker,
  create_sql_editor_connector
- sql_editor_dialog.py: 3997 -> 2608줄 (SQLEditorDialog + 전체 재수입)

원래 계획에서 LARGE_SQL_RENDER_LIMIT_BYTES와 create_sql_editor_connector는
파사드에 남기려 했으나, 실제로는 code_editor/workers 내부에서 참조해 순환
import가 발생함을 검증 중 발견 — 두 정의를 실제 사용처(각각 code_editor,
workers)로 옮기고 파사드에서 재수입하도록 수정. 워커가 create_sql_editor_
connector를 호출하는 시점의 이름 해석은 워커 자신의 모듈 네임스페이스를
따르므로, 관련 테스트의 monkeypatch 대상도 sql_editor_workers 모듈로 이동.

main_window.py/dialogs/__init__.py는 변경 없이 그대로 동작 (파사드 재수입).

검증: 전체 pytest 1809 passed / 1 known-flaky(macOS CI 의존) 제외 통과.
py_compile 전체 통과. 앱 스모크(core_hello) 성공.
