+++ src/ui/dialogs/sql_editor_dialog.py
@@ -20,39 +20,38 @@ from PyQt6.QtWidgets import (
 )
 from PyQt6.QtCore import Qt, QRect, QSize, pyqtSignal, QThread, QTimer, QPoint
 from PyQt6.QtGui import (
-    QSyntaxHighlighter, QTextCharFormat, QColor, QFont, QPainter,
+    QTextCharFormat, QColor, QPainter,
     QTextCursor, QKeySequence, QShortcut, QPen, QTextFormat
 )
 import re
 from typing import List, Dict, Optional, Tuple
 
-from src.core.db_core_service import create_rust_db_connector, normalize_db_engine
+from src.core.db_core_service import normalize_db_engine
 from src.core.sql_query_classifier import (
     classify_sql_statement,
     is_mysql_implicit_commit_ddl,
-    statement_returns_rows,
 )
 from src.core.sql_statement_parser import (
     find_sql_statement_at_position,
     parse_sql_statements,
 )
+from src.ui.dialogs.sql_editor_highlighters import SQLHighlighter, SQLValidatorHighlighter
+from src.ui.dialogs.sql_editor_autocomplete import AutoCompletePopup
+from src.ui.dialogs.sql_editor_code_editor import (
+    LineNumberArea,
+    CodeEditor,
+    ValidatingCodeEditor,
+    SQLEditorTab,
+    LARGE_SQL_RENDER_LIMIT_BYTES,
+)
+from src.ui.dialogs.sql_editor_history_dialog import HistoryDialog
+from src.ui.dialogs.sql_editor_workers import (
+    SQLQueryWorker,
+    SQLTransactionExecutionWorker,
+    create_sql_editor_connector,
+)
 
 logger = logging.getLogger(__name__)
-
-LARGE_SQL_RENDER_LIMIT_BYTES = 512 * 1024
-
-
-def create_sql_editor_connector(engine, host, port, user, password, database=None, schema=None):
-    db_engine = normalize_db_engine(engine, port)
-    return create_rust_db_connector(
-        db_engine,
-        host,
-        port,
-        user,
-        password,
-        database,
-        schema=(schema or "") if db_engine == "postgresql" else "",
-    )
 
 
 def format_metadata_db_version(db_version) -> str:
@@ -73,1439 +72,51 @@ def format_metadata_db_version(db_version) -> str:
 # =====================================================================
 # SQL 구문 하이라이터
 # =====================================================================
-class SQLHighlighter(QSyntaxHighlighter):
-    """SQL 구문 하이라이팅"""
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._init_rules()
-
-    def _init_rules(self):
-        """하이라이팅 규칙 초기화"""
-        self.highlighting_rules = []
-
-        # 키워드 포맷
-        keyword_format = QTextCharFormat()
-        keyword_format.setForeground(QColor("#569CD6"))  # 파란색
-        keyword_format.setFontWeight(QFont.Weight.Bold)
-
-        keywords = [
-            "SELECT", "FROM", "WHERE", "AND", "OR", "NOT", "IN", "LIKE",
-            "INSERT", "INTO", "VALUES", "UPDATE", "SET", "DELETE",
-            "CREATE", "ALTER", "DROP", "TABLE", "INDEX", "VIEW", "DATABASE",
-            "JOIN", "INNER", "LEFT", "RIGHT", "OUTER", "FULL", "CROSS", "ON",
-            "GROUP", "BY", "ORDER", "ASC", "DESC", "HAVING", "LIMIT", "OFFSET",
-            "UNION", "ALL", "DISTINCT", "AS", "CASE", "WHEN", "THEN", "ELSE", "END",
-            "NULL", "IS", "BETWEEN", "EXISTS", "PRIMARY", "KEY", "FOREIGN",
-            "REFERENCES", "CONSTRAINT", "DEFAULT", "AUTO_INCREMENT",
-            "TRUNCATE", "BEGIN", "COMMIT", "ROLLBACK", "TRANSACTION",
-            "IF", "ELSE", "WHILE", "DECLARE", "CURSOR", "FETCH", "PROCEDURE", "FUNCTION",
-            "RETURNS", "RETURN", "CALL", "TRIGGER", "BEFORE", "AFTER", "FOR", "EACH", "ROW",
-            "TRUE", "FALSE", "USE", "SHOW", "DESCRIBE", "EXPLAIN", "GRANT", "REVOKE"
-        ]
-
-        for word in keywords:
-            pattern = rf"\b{word}\b"
-            self.highlighting_rules.append((re.compile(pattern, re.IGNORECASE), keyword_format))
-
-        # 함수 포맷
-        function_format = QTextCharFormat()
-        function_format.setForeground(QColor("#DCDCAA"))  # 노란색
-
-        functions = [
-            "COUNT", "SUM", "AVG", "MIN", "MAX", "COALESCE", "IFNULL", "NULLIF",
-            "CONCAT", "SUBSTRING", "LENGTH", "TRIM", "UPPER", "LOWER", "REPLACE",
-            "NOW", "DATE", "TIME", "DATETIME", "TIMESTAMP", "YEAR", "MONTH", "DAY",
-            "HOUR", "MINUTE", "SECOND", "DATEDIFF", "DATE_ADD", "DATE_SUB",
-            "CAST", "CONVERT", "ROUND", "FLOOR", "CEIL", "ABS", "MOD", "POWER",
-            "GROUP_CONCAT", "JSON_EXTRACT", "JSON_ARRAY", "JSON_OBJECT"
-        ]
-
-        for word in functions:
-            pattern = rf"\b{word}\s*\("
-            self.highlighting_rules.append((re.compile(pattern, re.IGNORECASE), function_format))
-
-        # 숫자 포맷
-        number_format = QTextCharFormat()
-        number_format.setForeground(QColor("#B5CEA8"))  # 연두색
-        self.highlighting_rules.append((re.compile(r"\b\d+\.?\d*\b"), number_format))
-
-        # 문자열 포맷 (작은따옴표)
-        string_format = QTextCharFormat()
-        string_format.setForeground(QColor("#CE9178"))  # 주황색
-        self.highlighting_rules.append((re.compile(r"'[^']*'"), string_format))
-        self.highlighting_rules.append((re.compile(r'"[^"]*"'), string_format))
-
-        # 주석 포맷
-        comment_format = QTextCharFormat()
-        comment_format.setForeground(QColor("#6A9955"))  # 녹색
-        comment_format.setFontItalic(True)
-        self.highlighting_rules.append((re.compile(r"--[^\n]*"), comment_format))
-        self.highlighting_rules.append((re.compile(r"#[^\n]*"), comment_format))
-
-        # 멀티라인 주석 저장
-        self.multiline_comment_format = comment_format
-
-    def highlightBlock(self, text):
-        """블록 하이라이팅"""
-        # 일반 규칙 적용
-        for pattern, format_ in self.highlighting_rules:
-            for match in pattern.finditer(text):
-                start = match.start()
-                length = match.end() - start
-                self.setFormat(start, length, format_)
-
-        # 멀티라인 주석 처리
-        self.setCurrentBlockState(0)
-
-        start_index = 0
-        if self.previousBlockState() != 1:
-            start_match = re.search(r"/\*", text)
-            start_index = start_match.start() if start_match else -1
-
-        while start_index >= 0:
-            end_match = re.search(r"\*/", text[start_index + 2:])
-            if end_match:
-                end_index = start_index + 2 + end_match.end()
-                comment_length = end_index - start_index
-            else:
-                self.setCurrentBlockState(1)
-                comment_length = len(text) - start_index
-
-            self.setFormat(start_index, comment_length, self.multiline_comment_format)
-
-            start_match = re.search(r"/\*", text[start_index + comment_length:])
-            start_index = (start_index + comment_length + start_match.start()) if start_match else -1
 
 
 # =====================================================================
 # SQL 검증 하이라이터 (밑줄 표시)
 # =====================================================================
-class SQLValidatorHighlighter(SQLHighlighter):
-    """SQL 구문 하이라이터 + 검증 이슈 밑줄 표시"""
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self._issues = []  # ValidationIssue 목록
-        self._issue_formats = {}  # line -> [(col, end_col, format), ...]
-
-        # 에러 포맷 (빨간 물결 밑줄)
-        self.error_format = QTextCharFormat()
-        self.error_format.setUnderlineStyle(QTextCharFormat.UnderlineStyle.WaveUnderline)
-        self.error_format.setUnderlineColor(QColor("#E74C3C"))  # 빨간색
-
-        # 경고 포맷 (노란 물결 밑줄)
-        self.warning_format = QTextCharFormat()
-        self.warning_format.setUnderlineStyle(QTextCharFormat.UnderlineStyle.WaveUnderline)
-        self.warning_format.setUnderlineColor(QColor("#F39C12"))  # 노란색
-
-        # 정보 포맷 (파란 밑줄)
-        self.info_format = QTextCharFormat()
-        self.info_format.setUnderlineStyle(QTextCharFormat.UnderlineStyle.SingleUnderline)
-        self.info_format.setUnderlineColor(QColor("#3498DB"))  # 파란색
-
-    def set_issues(self, issues: list):
-        """검증 이슈 설정 및 재하이라이팅"""
-        self._issues = issues
-        self._build_issue_map()
-        self.rehighlight()
-
-    def _build_issue_map(self):
-        """줄별 이슈 맵 생성"""
-        self._issue_formats = {}
-
-        for issue in self._issues:
-            line = issue.line
-            if line not in self._issue_formats:
-                self._issue_formats[line] = []
-
-            # 심각도에 따른 포맷 선택
-            from src.core.sql_validator import IssueSeverity
-            if issue.severity == IssueSeverity.ERROR:
-                fmt = self.error_format
-            elif issue.severity == IssueSeverity.WARNING:
-                fmt = self.warning_format
-            else:
-                fmt = self.info_format
-
-            self._issue_formats[line].append((issue.column, issue.end_column, fmt))
-
-    def highlightBlock(self, text):
-        """블록 하이라이팅 (기본 + 검증 이슈)"""
-        # 기본 SQL 하이라이팅
-        super().highlightBlock(text)
-
-        # 검증 이슈 밑줄 추가
-        block_number = self.currentBlock().blockNumber()
-        if block_number in self._issue_formats:
-            for col, end_col, fmt in self._issue_formats[block_number]:
-                # 범위 검증
-                start = max(0, col)
-                length = min(end_col, len(text)) - start
-                if length > 0:
-                    self.setFormat(start, length, fmt)
-
-    def get_issues(self) -> list:
-        """현재 이슈 목록 반환"""
-        return self._issues
 
 
 # =====================================================================
 # 자동완성 팝업
 # =====================================================================
-class AutoCompletePopup(QListWidget):
-    """SQL 자동완성 팝업 위젯"""
-
-    item_selected = pyqtSignal(str)  # 선택된 텍스트
-
-    # 항목당 높이 (padding 포함)
-    ITEM_HEIGHT = 24
-    MAX_HEIGHT = 200
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        # Tool 윈도우로 설정하여 포커스를 부모에게 유지
-        self.setWindowFlags(
-            Qt.WindowType.Tool |
-            Qt.WindowType.FramelessWindowHint |
-            Qt.WindowType.WindowStaysOnTopHint
-        )
-        self.setFocusPolicy(Qt.FocusPolicy.NoFocus)
-        self.setAttribute(Qt.WidgetAttribute.WA_ShowWithoutActivating)  # 포커스 가져가지 않음
-        self.setMouseTracking(True)
-        self.setMinimumWidth(200)
-
-        self.setStyleSheet("""
-            QListWidget {
-                background-color: #2D2D2D;
-                color: #D4D4D4;
-                border: 1px solid #454545;
-                font-family: 'Consolas', monospace;
-                font-size: 12px;
-            }
-            QListWidget::item {
-                padding: 4px 8px;
-            }
-            QListWidget::item:selected {
-                background-color: #094771;
-                color: white;
-            }
-            QListWidget::item:hover {
-                background-color: #383838;
-            }
-        """)
-
-        self.itemClicked.connect(self._on_item_clicked)
-        self.itemDoubleClicked.connect(self._on_item_clicked)
-
-    def set_completions(self, completions: List[Dict]):
-        """자동완성 목록 설정
-
-        Args:
-            completions: [{label, type, detail}, ...]
-        """
-        self.clear()
-
-        for item_data in completions:
-            label = item_data.get('label', '')
-            item_type = item_data.get('type', '')
-            detail = item_data.get('detail', '')
-
-            # 아이콘 접두사
-            icons = {
-                'table': '📋',
-                'column': '📊',
-                'keyword': '🔤',
-                'function': '⚡',
-            }
-            icon = icons.get(item_type, '•')
-
-            display_text = f"{icon} {label}"
-            if detail:
-                display_text += f"  ({detail})"
-
-            list_item = QListWidgetItem(display_text)
-            list_item.setData(Qt.ItemDataRole.UserRole, label)
-            self.addItem(list_item)
-
-        if self.count() > 0:
-            self.setCurrentRow(0)
-            self._adjust_height(self.count())
-
-    def filter_items(self, prefix: str):
-        """입력에 따라 항목 필터링"""
-        prefix_lower = prefix.lower()
-        visible_count = 0
-
-        for i in range(self.count()):
-            item = self.item(i)
-            label = item.data(Qt.ItemDataRole.UserRole) or ""
-            matches = label.lower().startswith(prefix_lower)
-            item.setHidden(not matches)
-            if matches:
-                visible_count += 1
-
-        # 첫 번째 보이는 항목 선택
-        for i in range(self.count()):
-            if not self.item(i).isHidden():
-                self.setCurrentRow(i)
-                break
-
-        # 높이 조절
-        self._adjust_height(visible_count)
-
-        return visible_count > 0
-
-    def _adjust_height(self, item_count: int):
-        """항목 개수에 따라 높이 조절"""
-        if item_count <= 0:
-            return
-
-        # 항목 개수 * 항목 높이 + 테두리 여백
-        calculated_height = item_count * self.ITEM_HEIGHT + 4
-        # 최대 높이 제한
-        new_height = min(calculated_height, self.MAX_HEIGHT)
-        self.setFixedHeight(new_height)
-
-    def _on_item_clicked(self, item):
-        """항목 클릭"""
-        label = item.data(Qt.ItemDataRole.UserRole)
-        if label:
-            self.item_selected.emit(label)
-            self.hide()
-
-    def select_current(self):
-        """현재 선택 항목 확정"""
-        item = self.currentItem()
-        if item and not item.isHidden():
-            label = item.data(Qt.ItemDataRole.UserRole)
-            if label:
-                self.item_selected.emit(label)
-                self.hide()
-                return True
-        return False
-
-    def move_selection(self, direction: int):
-        """선택 이동 (위/아래)"""
-        current = self.currentRow()
-        new_row = current + direction
-
-        # 보이는 항목만 선택
-        while 0 <= new_row < self.count():
-            if not self.item(new_row).isHidden():
-                self.setCurrentRow(new_row)
-                return
-            new_row += direction
-
-    def keyPressEvent(self, event):
-        """키 이벤트 (팝업 내부) - 부모 에디터로 전달"""
-        if event.key() == Qt.Key.Key_Escape:
-            self.hide()
-        elif event.key() in (Qt.Key.Key_Return, Qt.Key.Key_Enter, Qt.Key.Key_Tab):
-            self.select_current()
-        elif event.key() == Qt.Key.Key_Up:
-            self.move_selection(-1)
-        elif event.key() == Qt.Key.Key_Down:
-            self.move_selection(1)
-        else:
-            # 다른 키 입력은 부모 에디터로 전달
-            if self.parent():
-                self.parent().keyPressEvent(event)
 
 
 # =====================================================================
 # 줄 번호 위젯
 # =====================================================================
-class LineNumberArea(QWidget):
-    """줄 번호 표시 영역"""
-
-    def __init__(self, editor):
-        super().__init__(editor)
-        self.code_editor = editor
-
-    def sizeHint(self):
-        return QSize(self.code_editor.line_number_area_width(), 0)
-
-    def paintEvent(self, event):
-        self.code_editor.line_number_area_paint_event(event)
 
 
 # =====================================================================
 # 코드 에디터 (줄 번호 + 하이라이팅)
 # =====================================================================
-class CodeEditor(QPlainTextEdit):
-    """줄 번호가 있는 코드 에디터"""
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        # 에디터 스타일
-        self.setStyleSheet("""
-            QPlainTextEdit {
-                background-color: #1E1E1E;
-                color: #D4D4D4;
-                font-family: 'Consolas', 'Courier New', monospace;
-                font-size: 13px;
-                border: 1px solid #3C3C3C;
-                selection-background-color: #264F78;
-            }
-        """)
-
-        # 줄 번호 영역
-        self.line_number_area = LineNumberArea(self)
-
-        # 신호 연결
-        self.blockCountChanged.connect(self.update_line_number_area_width)
-        self.updateRequest.connect(self.update_line_number_area)
-        self.cursorPositionChanged.connect(self.highlight_current_line)
-
-        self.update_line_number_area_width(0)
-        self.highlight_current_line()
-
-        # SQL 하이라이터
-        self.highlighter = SQLHighlighter(self.document())
-
-        # 탭 크기 설정
-        self.setTabStopDistance(40)
-
-    def line_number_area_width(self):
-        """줄 번호 영역 너비 계산"""
-        digits = 1
-        max_num = max(1, self.blockCount())
-        while max_num >= 10:
-            max_num //= 10
-            digits += 1
-        space = 10 + self.fontMetrics().horizontalAdvance('9') * digits
-        return space
-
-    def update_line_number_area_width(self, _):
-        """줄 번호 영역 너비 업데이트"""
-        self.setViewportMargins(self.line_number_area_width(), 0, 0, 0)
-
-    def update_line_number_area(self, rect, dy):
-        """줄 번호 영역 업데이트"""
-        if dy:
-            self.line_number_area.scroll(0, dy)
-        else:
-            self.line_number_area.update(0, rect.y(), self.line_number_area.width(), rect.height())
-
-        if rect.contains(self.viewport().rect()):
-            self.update_line_number_area_width(0)
-
-    def resizeEvent(self, event):
-        """리사이즈 이벤트"""
-        super().resizeEvent(event)
-        cr = self.contentsRect()
-        self.line_number_area.setGeometry(QRect(cr.left(), cr.top(), self.line_number_area_width(), cr.height()))
-
-    def line_number_area_paint_event(self, event):
-        """줄 번호 그리기"""
-        painter = QPainter(self.line_number_area)
-        painter.fillRect(event.rect(), QColor("#252526"))
-
-        block = self.firstVisibleBlock()
-        block_number = block.blockNumber()
-        top = int(self.blockBoundingGeometry(block).translated(self.contentOffset()).top())
-        bottom = top + int(self.blockBoundingRect(block).height())
-
-        while block.isValid() and top <= event.rect().bottom():
-            if block.isVisible() and bottom >= event.rect().top():
-                number = str(block_number + 1)
-                painter.setPen(QColor("#858585"))
-                painter.drawText(0, top, self.line_number_area.width() - 5,
-                                 self.fontMetrics().height(), Qt.AlignmentFlag.AlignRight, number)
-
-            block = block.next()
-            top = bottom
-            bottom = top + int(self.blockBoundingRect(block).height())
-            block_number += 1
-
-    def highlight_current_line(self):
-        """현재 줄 하이라이트"""
-        extra_selections = []
-
-        if not self.isReadOnly():
-            selection = QTextEdit.ExtraSelection()
-            line_color = QColor("#2D2D2D")
-            selection.format.setBackground(line_color)
-            selection.format.setProperty(QTextCharFormat.Property.FullWidthSelection, True)
-            selection.cursor = self.textCursor()
-            selection.cursor.clearSelection()
-            extra_selections.append(selection)
-
-        self.setExtraSelections(extra_selections)
 
 
 # =====================================================================
 # 검증 기능이 있는 코드 에디터
 # =====================================================================
-class ValidatingCodeEditor(CodeEditor):
-    """검증 + 자동완성 기능이 있는 코드 에디터"""
-
-    validation_requested = pyqtSignal(str)  # SQL 텍스트
-    autocomplete_requested = pyqtSignal(str, int)  # SQL, cursor_pos
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-
-        # 검증 하이라이터로 교체
-        self.highlighter = SQLValidatorHighlighter(self.document())
-        self._large_document_mode = False
-
-        # 검증 이슈 목록 (툴팁용)
-        self._issues: List = []
-
-        # Debounce 타이머 (300ms)
-        self._validation_timer = QTimer(self)
-        self._validation_timer.setSingleShot(True)
-        self._validation_timer.setInterval(300)
-        self._validation_timer.timeout.connect(self._trigger_validation)
-
-        # 자동완성 팝업
-        self._autocomplete_popup = AutoCompletePopup(self)
-        self._autocomplete_popup.item_selected.connect(self._on_autocomplete_selected)
-        self._autocomplete_prefix = ""  # 현재 입력 중인 접두사
-
-        # 마우스 이동 추적 (호버 툴팁)
-        self.setMouseTracking(True)
-
-        # 텍스트 변경 시 검증 예약
-        self.textChanged.connect(self._schedule_validation)
-
-    def _schedule_validation(self):
-        """검증 예약 (debounce)"""
-        if self._large_document_mode:
-            return
-        self._validation_timer.stop()
-        self._validation_timer.start()
-
-    def _trigger_validation(self):
-        """검증 실행 시그널 발생"""
-        if self._large_document_mode:
-            return
-        self.validation_requested.emit(self.toPlainText())
-
-    def set_validation_issues(self, issues: list):
-        """검증 이슈 설정"""
-        self._issues = issues
-        if not self._large_document_mode:
-            self.highlighter.set_issues(issues)
-
-    def set_large_document_mode(self, enabled: bool):
-        """Disable expensive whole-document features for large SQL text."""
-        enabled = bool(enabled)
-        if self._large_document_mode == enabled:
-            return
-
-        self._large_document_mode = enabled
-        self._validation_timer.stop()
-        self._issues = []
-
-        if enabled:
-            if self.highlighter:
-                self.highlighter.setDocument(None)
-            self.setToolTip("대용량 SQL 파일: 구문 하이라이트와 실시간 검증이 비활성화되었습니다.")
-        else:
-            self.highlighter = SQLValidatorHighlighter(self.document())
-            self.setToolTip("")
-
-    def is_large_document_mode(self) -> bool:
-        return self._large_document_mode
-
-    def get_issue_at_position(self, pos: int) -> Optional[object]:
-        """특정 위치의 이슈 반환"""
-        # 위치 → 줄/컬럼 변환
-        block = self.document().findBlock(pos)
-        line = block.blockNumber()
-        col = pos - block.position()
-
-        for issue in self._issues:
-            if issue.line == line and issue.column <= col < issue.end_column:
-                return issue
-
-        return None
-
-    def mouseMoveEvent(self, event):
-        """마우스 이동 시 툴팁 표시"""
-        super().mouseMoveEvent(event)
-
-        # 커서 위치에서 문자 위치 계산
-        cursor = self.cursorForPosition(event.pos())
-        pos = cursor.position()
-
-        issue = self.get_issue_at_position(pos)
-        if issue:
-            # 툴팁 내용 생성
-            tooltip_lines = [issue.message]
-            if issue.suggestions:
-                tooltip_lines.append(f"💡 제안: {', '.join(issue.suggestions)}")
-
-            QToolTip.showText(event.globalPosition().toPoint(), '\n'.join(tooltip_lines), self)
-        else:
-            QToolTip.hideText()
-
-    def keyPressEvent(self, event):
-        """키 입력 처리 (자동완성 포함)"""
-        # 자동완성 팝업이 열려있을 때
-        if self._autocomplete_popup.isVisible():
-            if event.key() == Qt.Key.Key_Escape:
-                self._autocomplete_popup.hide()
-                return
-            elif event.key() in (Qt.Key.Key_Return, Qt.Key.Key_Enter, Qt.Key.Key_Tab):
-                if self._autocomplete_popup.select_current():
-                    return
-                # 선택 실패 시 (빈 목록 등) 팝업 닫고 기본 동작
-                self._autocomplete_popup.hide()
-            elif event.key() == Qt.Key.Key_Up:
-                self._autocomplete_popup.move_selection(-1)
-                return
-            elif event.key() == Qt.Key.Key_Down:
-                self._autocomplete_popup.move_selection(1)
-                return
-            # 그 외 키 입력은 에디터에 전달 후 필터링 업데이트
-
-        # Ctrl+Space: 자동완성 표시
-        if event.modifiers() == Qt.KeyboardModifier.ControlModifier and event.key() == Qt.Key.Key_Space:
-            self._show_autocomplete()
-            return
-
-        # '.' 입력 후 자동완성 (table. 패턴)
-        if event.text() == '.':
-            super().keyPressEvent(event)
-            # 잠시 후 자동완성 표시
-            QTimer.singleShot(50, self._show_autocomplete)
-            return
-
-        # 일반 키 입력 처리
-        super().keyPressEvent(event)
-
-        # 팝업이 열려있으면 필터링 업데이트
-        if self._autocomplete_popup.isVisible():
-            # 공백이나 특수문자 입력 시 팝업 닫기
-            if event.text() and not event.text().isalnum() and event.text() != '_':
-                self._autocomplete_popup.hide()
-            else:
-                self._update_autocomplete_filter()
-
-    def _show_autocomplete(self):
-        """자동완성 팝업 표시 요청"""
-        cursor = self.textCursor()
-        pos = cursor.position()
-        self._autocomplete_prefix = self._get_current_word()
-        self.autocomplete_requested.emit(self.toPlainText(), pos)
-
-    def show_autocomplete_popup(self, completions: List[Dict]):
-        """자동완성 팝업 표시"""
-        if not completions:
-            self._autocomplete_popup.hide()
-            return
-
-        self._autocomplete_popup.set_completions(completions)
-
-        # 필터링
-        if self._autocomplete_prefix:
-            if not self._autocomplete_popup.filter_items(self._autocomplete_prefix):
-                self._autocomplete_popup.hide()
-                return
-
-        # 팝업 위치 계산 (커서 아래)
-        cursor_rect = self.cursorRect()
-        global_pos = self.mapToGlobal(cursor_rect.bottomLeft())
-        self._autocomplete_popup.move(global_pos)
-        self._autocomplete_popup.show()
-
-        # 에디터에 포커스 유지 (팝업이 포커스 가져가지 않도록)
-        self.setFocus()
-        self.activateWindow()
-
-    def _update_autocomplete_filter(self):
-        """자동완성 필터 업데이트"""
-        self._autocomplete_prefix = self._get_current_word()
-        if not self._autocomplete_popup.filter_items(self._autocomplete_prefix):
-            self._autocomplete_popup.hide()
-
-    def _get_current_word(self) -> str:
-        """커서 위치의 현재 단어 추출"""
-        cursor = self.textCursor()
-        cursor.movePosition(QTextCursor.MoveOperation.StartOfWord, QTextCursor.MoveMode.KeepAnchor)
-        return cursor.selectedText()
-
-    def _on_autocomplete_selected(self, text: str):
-        """자동완성 항목 선택됨"""
-        cursor = self.textCursor()
-
-        # 현재 단어 선택 후 교체
-        cursor.movePosition(QTextCursor.MoveOperation.StartOfWord, QTextCursor.MoveMode.KeepAnchor)
-        cursor.insertText(text)
-
-        self.setTextCursor(cursor)
-
-        # 에디터로 포커스 복원
-        self.setFocus()
-        self.activateWindow()
-
-    def focusOutEvent(self, event):
-        """포커스 잃을 때 자동완성 숨김"""
-        # 자동완성 팝업으로 포커스 이동 시 숨기지 않음
-        if self._autocomplete_popup.isVisible():
-            # 약간의 딜레이 후 숨김 (팝업 클릭 허용)
-            QTimer.singleShot(100, self._maybe_hide_autocomplete)
-        super().focusOutEvent(event)
-
-    def _maybe_hide_autocomplete(self):
-        """자동완성 팝업 숨김 (조건부)"""
-        if not self.hasFocus() and not self._autocomplete_popup.underMouse():
-            self._autocomplete_popup.hide()
 
 
 # =====================================================================
 # SQL 쿼리 실행 워커 (자동 커밋 모드)
 # =====================================================================
-class SQLQueryWorker(QThread):
-    """SQL 쿼리 실행 워커 (자동 커밋)"""
-    progress = pyqtSignal(str)
-    query_result = pyqtSignal(int, bool, list, list, str, int, float)  # idx, returns_rows, columns, rows, error, affected, time
-    finished = pyqtSignal(bool, str)
-
-    def __init__(self, host, port, user, password, database, queries, engine="mysql", schema=None):
-        super().__init__()
-        self.engine = normalize_db_engine(engine, port)
-        self.host = host
-        self.port = port
-        self.user = user
-        self.password = password
-        self.database = database
-        self.schema = schema
-        self.queries = queries  # List of query strings
-
-    def run(self):
-        connector = None
-        try:
-            connector = create_sql_editor_connector(
-                self.engine,
-                self.host,
-                self.port,
-                self.user,
-                self.password,
-                self.database,
-                self.schema,
-            )
-            success, msg = connector.connect()
-
-            if not success:
-                self.finished.emit(False, f"연결 실패: {msg}")
-                return
-
-            self.progress.emit(f"✅ 연결 성공: {self.host}:{self.port}")
-            connector.connection.autocommit(True)
-
-            total_queries = len(self.queries)
-            success_count = 0
-            error_count = 0
-
-            for idx, query in enumerate(self.queries):
-                if self.isInterruptionRequested():
-                    self.finished.emit(False, "⚠️ 실행이 취소되었습니다")
-                    return
-
-                query = query.strip()
-                if not query:
-                    continue
-
-                self.progress.emit(f"📄 쿼리 {idx + 1}/{total_queries} 실행 중...")
-
-                start_time = time.time()
-                try:
-                    if statement_returns_rows(query):
-                        rows = []
-
-                        def collect_batch(batch):
-                            rows.extend(batch)
-
-                        result = connector.connection.facade.execute_on_connection_streaming(
-                            connector.connection.connection_id,
-                            query,
-                            row_batch_size=500,
-                            on_batch=collect_batch,
-                        )
-                        columns = result.get("columns") or []
-                        row_list = [[row.get(col) for col in columns] for row in rows]
-                        execution_time = time.time() - start_time
-                        self.query_result.emit(idx, True, columns, row_list, "", len(row_list), execution_time)
-                        success_count += 1
-                        continue
-
-                    # 직접 커서 사용하여 실행
-                    with connector.connection.cursor() as cursor:
-                        cursor.execute(query)
-
-                        # 행을 반환하는 statement인지 확인 (None만 비행-statement)
-                        if cursor.description is not None:
-                            # SELECT 결과 (0행이어도 columns == [] 로 반환됨)
-                            columns = [desc[0] for desc in cursor.description]
-                            rows = cursor.fetchall()
-                            # Dict to list 변환
-                            row_list = []
-                            for row in rows:
-                                if isinstance(row, dict):
-                                    row_list.append([row.get(col) for col in columns])
-                                else:
-                                    row_list.append(list(row))
-
-                            execution_time = time.time() - start_time
-                            self.query_result.emit(idx, True, columns, row_list, "", len(row_list), execution_time)
-                            success_count += 1
-                        else:
-                            # INSERT, UPDATE, DELETE 등
-                            affected = cursor.rowcount
-                            connector.connection.commit()
-                            execution_time = time.time() - start_time
-                            self.query_result.emit(idx, False, [], [], "", affected, execution_time)
-                            success_count += 1
-
-                except Exception as e:
-                    execution_time = time.time() - start_time
-                    self.query_result.emit(
-                        idx, statement_returns_rows(query), [], [], str(e), 0, execution_time
-                    )
-                    error_count += 1
-
-            if error_count == 0:
-                self.finished.emit(True, f"✅ {success_count}개 쿼리 실행 완료")
-            else:
-                self.finished.emit(False, f"⚠️ {success_count}개 성공, {error_count}개 실패")
-
-        except Exception as e:
-            self.finished.emit(False, f"❌ 오류: {str(e)}")
-
-        finally:
-            # 연결 정리
-            if connector:
-                try:
-                    connector.disconnect()
-                except Exception:
-                    logger.debug("자동 커밋 워커 연결 정리 실패", exc_info=True)
 
 
 # =====================================================================
 # SQL 트랜잭션 실행 워커 (지속 연결 모드 - 커밋/롤백은 다이얼로그가 소유)
 # =====================================================================
-class SQLTransactionExecutionWorker(QThread):
-    """지속 트랜잭션 연결에서 쿼리를 순차 실행하는 워커.
-
-    커밋/롤백은 이 워커가 아니라 SQLEditorDialog가 소유한 연결에서 처리한다.
-    PostgreSQL은 에러 발생 시 트랜잭션 전체가 aborted 상태가 되므로 즉시 롤백하고 중단한다.
-    """
-    progress = pyqtSignal(int, int, str, str)  # idx, total, query_type, preview
-    query_result = pyqtSignal(int, str, bool, list, list, str, int, float)  # idx, query, returns_rows, columns, rows, error, affected, time
-    postgres_rolled_back = pyqtSignal(str)
-    finished = pyqtSignal(bool, str)
-
-    def __init__(self, connection, queries, engine):
-        super().__init__()
-        self.connection = connection
-        self.queries = queries
-        self.engine = engine
-
-    def run(self):
-        total = len(self.queries)
-        for idx, raw_query in enumerate(self.queries):
-            if self.isInterruptionRequested():
-                self.finished.emit(False, "⚠️ 실행이 취소되었습니다")
-                return
-
-            query = raw_query.strip()
-            if not query:
-                continue
-
-            classification = classify_sql_statement(query)
-            query_type = (classification.leading_keyword or "other").upper()
-            preview = query[:100] + ("..." if len(query) > 100 else "")
-            self.progress.emit(idx, total, query_type, preview)
-
-            start_time = time.time()
-            try:
-                with self.connection.cursor() as cursor:
-                    cursor.execute(query)
-
-                    if cursor.description is not None:
-                        columns = [desc[0] for desc in cursor.description]
-                        rows = cursor.fetchall()
-                        row_list = []
-                        for row in rows:
-                            if isinstance(row, dict):
-                                row_list.append([row.get(col) for col in columns])
-                            else:
-                                row_list.append(list(row))
-                        execution_time = time.time() - start_time
-                        self.query_result.emit(idx, query, True, columns, row_list, "", len(row_list), execution_time)
-                    else:
-                        affected = cursor.rowcount
-                        execution_time = time.time() - start_time
-                        self.query_result.emit(idx, query, False, [], [], "", affected, execution_time)
-
-            except Exception as e:
-                execution_time = time.time() - start_time
-                if self.engine == "postgresql":
-                    try:
-                        self.connection.rollback()
-                    except Exception:
-                        logger.debug("PostgreSQL 오류 후 롤백 실패", exc_info=True)
-                    self.postgres_rolled_back.emit(str(e))
-                    self.query_result.emit(idx, query, False, [], [], str(e), 0, execution_time)
-                    self.finished.emit(False, "❌ PostgreSQL 오류로 트랜잭션이 롤백되었습니다")
-                    return
-                # MySQL 등: 이전 쿼리는 이미 반영되었으므로 실패만 기록하고 계속 진행
-                self.query_result.emit(idx, query, False, [], [], str(e), 0, execution_time)
-
-        self.finished.emit(True, "✅ 실행 완료")
 
 
 # =====================================================================
 # 히스토리 다이얼로그
 # =====================================================================
-class HistoryDialog(QDialog):
-    """쿼리 히스토리 다이얼로그 (영구 보관, 고급 검색, 즐겨찾기)"""
-    query_selected = pyqtSignal(str)
-
-    ITEMS_PER_PAGE = 50  # 한 번에 로드할 항목 수
-
-    def __init__(self, parent, history_manager):
-        super().__init__(parent)
-        self.history_manager = history_manager
-        self.current_offset = 0
-        self.total_count = 0
-        self._history_items = []  # 현재 표시된 항목 데이터
-        self._is_searching = False  # 검색 모드 여부
-        self.setWindowTitle("쿼리 히스토리")
-        self.setMinimumSize(800, 600)
-        self.init_ui()
-        self.load_history()
-
-    def init_ui(self):
-        layout = QVBoxLayout(self)
-
-        # === 검색 필터 패널 ===
-        filter_group = QGroupBox("검색 필터")
-        filter_layout = QVBoxLayout(filter_group)
-
-        # 첫 번째 줄: 키워드, 날짜
-        row1 = QHBoxLayout()
-
-        row1.addWidget(QLabel("키워드:"))
-        self.keyword_edit = QLineEdit()
-        self.keyword_edit.setPlaceholderText("쿼리 내용으로 검색...")
-        self.keyword_edit.setMinimumWidth(200)
-        self.keyword_edit.returnPressed.connect(self._do_search)
-        row1.addWidget(self.keyword_edit)
-
-        row1.addWidget(QLabel("기간:"))
-
-        from PyQt6.QtWidgets import QDateEdit
-        from PyQt6.QtCore import QDate
-        self.date_from = QDateEdit()
-        self.date_from.setCalendarPopup(True)
-        self.date_from.setDate(QDate.currentDate().addMonths(-1))
-        self.date_from.setDisplayFormat("yyyy-MM-dd")
-        row1.addWidget(self.date_from)
-
-        row1.addWidget(QLabel("~"))
-
-        self.date_to = QDateEdit()
-        self.date_to.setCalendarPopup(True)
-        self.date_to.setDate(QDate.currentDate())
-        self.date_to.setDisplayFormat("yyyy-MM-dd")
-        row1.addWidget(self.date_to)
-
-        self.date_filter_check = QCheckBox("날짜 적용")
-        self.date_filter_check.setChecked(False)
-        row1.addWidget(self.date_filter_check)
-
-        row1.addStretch()
-        filter_layout.addLayout(row1)
-
-        # 두 번째 줄: 체크박스들, 버튼
-        row2 = QHBoxLayout()
-
-        self.success_check = QCheckBox("성공만")
-        row2.addWidget(self.success_check)
-
-        self.failure_check = QCheckBox("실패만")
-        row2.addWidget(self.failure_check)
-
-        self.favorites_check = QCheckBox("즐겨찾기만")
-        row2.addWidget(self.favorites_check)
-
-        row2.addStretch()
-
-        btn_search = QPushButton("🔍 검색")
-        btn_search.setStyleSheet("""
-            QPushButton {
-                background-color: #3498db; color: white;
-                padding: 6px 16px; border-radius: 4px;
-            }
-            QPushButton:hover { background-color: #2980b9; }
-        """)
-        btn_search.clicked.connect(self._do_search)
-        row2.addWidget(btn_search)
-
-        btn_reset = QPushButton("초기화")
-        btn_reset.setStyleSheet("""
-            QPushButton {
-                background-color: #95a5a6; color: white;
-                padding: 6px 12px; border-radius: 4px;
-            }
-            QPushButton:hover { background-color: #7f8c8d; }
-        """)
-        btn_reset.clicked.connect(self._reset_search)
-        row2.addWidget(btn_reset)
-
-        filter_layout.addLayout(row2)
-        layout.addWidget(filter_group)
-
-        # === 정보 바 ===
-        info_layout = QHBoxLayout()
-        self.info_label = QLabel("히스토리 로딩 중...")
-        self.info_label.setStyleSheet("color: #666;")
-        info_layout.addWidget(self.info_label)
-
-        self.fav_count_label = QLabel("⭐ 0")
-        self.fav_count_label.setStyleSheet("color: #f39c12; font-weight: bold;")
-        info_layout.addWidget(self.fav_count_label)
-
-        info_layout.addStretch()
-        layout.addLayout(info_layout)
-
-        # === 히스토리 리스트 ===
-        self.list_widget = QListWidget()
-        self.list_widget.setStyleSheet("""
-            QListWidget {
-                font-family: 'Consolas', monospace;
-                font-size: 12px;
-            }
-            QListWidget::item {
-                padding: 8px;
-                border-bottom: 1px solid #ddd;
-            }
-            QListWidget::item:selected {
-                background-color: #3498db;
-                color: white;
-            }
-        """)
-        self.list_widget.itemDoubleClicked.connect(self.select_query)
-        self.list_widget.setContextMenuPolicy(Qt.ContextMenuPolicy.CustomContextMenu)
-        self.list_widget.customContextMenuRequested.connect(self._show_context_menu)
-        layout.addWidget(self.list_widget)
-
-        # === 미리보기 ===
-        preview_group = QGroupBox("쿼리 미리보기")
-        preview_layout = QVBoxLayout(preview_group)
-        self.preview_text = QTextEdit()
-        self.preview_text.setReadOnly(True)
-        self.preview_text.setMaximumHeight(120)
-        self.preview_text.setStyleSheet("""
-            QTextEdit {
-                background-color: #2c3e50;
-                color: #ecf0f1;
-                font-family: 'Consolas', monospace;
-                font-size: 12px;
-            }
-        """)
-        preview_layout.addWidget(self.preview_text)
-        layout.addWidget(preview_group)
-
-        # 선택 시 미리보기 업데이트
-        self.list_widget.currentRowChanged.connect(self.update_preview)
-
-        # === 페이지네이션 ===
-        page_layout = QHBoxLayout()
-
-        self.btn_load_more = QPushButton("📜 더 보기")
-        self.btn_load_more.setStyleSheet("""
-            QPushButton {
-                background-color: #6c757d; color: white;
-                padding: 6px 16px; border-radius: 4px;
-            }
-            QPushButton:hover { background-color: #5a6268; }
-            QPushButton:disabled { background-color: #adb5bd; }
-        """)
-        self.btn_load_more.clicked.connect(self.load_more)
-        self.btn_load_more.setVisible(False)
-
-        page_layout.addStretch()
-        page_layout.addWidget(self.btn_load_more)
-        page_layout.addStretch()
-        layout.addLayout(page_layout)
-
-        # === 하단 버튼 ===
-        btn_layout = QHBoxLayout()
-
-        btn_fav_toggle = QPushButton("⭐ 즐겨찾기 토글")
-        btn_fav_toggle.setStyleSheet("""
-            QPushButton {
-                background-color: #f39c12; color: white;
-                padding: 8px 16px; border-radius: 4px;
-            }
-            QPushButton:hover { background-color: #d68910; }
-        """)
-        btn_fav_toggle.clicked.connect(self._toggle_favorite)
-        btn_layout.addWidget(btn_fav_toggle)
-
-        btn_layout.addStretch()
-
-        btn_use = QPushButton("📋 에디터에 붙여넣기")
-        btn_use.setStyleSheet("""
-            QPushButton {
-                background-color: #3498db; color: white; font-weight: bold;
-                padding: 8px 16px; border-radius: 4px;
-            }
-            QPushButton:hover { background-color: #2980b9; }
-        """)
-        btn_use.clicked.connect(self.select_current)
-        btn_layout.addWidget(btn_use)
-
-        btn_close = QPushButton("닫기")
-        btn_close.clicked.connect(self.reject)
-        btn_layout.addWidget(btn_close)
-
-        layout.addLayout(btn_layout)
-
-    def load_history(self):
-        """히스토리 초기 로드"""
-        self.list_widget.clear()
-        self._history_items.clear()
-        self.current_offset = 0
-        self._is_searching = False
-        self._load_chunk()
-        self._update_fav_count()
-
-    def _load_chunk(self):
-        """히스토리 청크 로드"""
-        if self._is_searching:
-            self._load_search_chunk()
-        else:
-            history, self.total_count = self.history_manager.get_history(
-                limit=self.ITEMS_PER_PAGE,
-                offset=self.current_offset
-            )
-
-            for item in history:
-                self._add_history_item(item)
-                self._history_items.append(item)
-
-            self.current_offset += len(history)
-
-        # 정보 라벨 업데이트
-        loaded = self.list_widget.count()
-        self.info_label.setText(f"📊 {loaded:,} / {self.total_count:,}개 표시")
-
-        # 더 보기 버튼 표시/숨김
-        has_more = self.current_offset < self.total_count
-        self.btn_load_more.setVisible(has_more)
-        if has_more:
-            remaining = self.total_count - self.current_offset
-            self.btn_load_more.setText(f"📜 더 보기 ({remaining:,}개 남음)")
-
-    def _load_search_chunk(self):
-        """검색 결과 청크 로드"""
-        from datetime import datetime
-
-        keyword = self.keyword_edit.text().strip() or None
-
-        date_from = None
-        date_to = None
-        if self.date_filter_check.isChecked():
-            date_from = datetime(
-                self.date_from.date().year(),
-                self.date_from.date().month(),
-                self.date_from.date().day()
-            )
-            date_to = datetime(
-                self.date_to.date().year(),
-                self.date_to.date().month(),
-                self.date_to.date().day()
-            )
-
-        success_only = None
-        if self.success_check.isChecked():
-            success_only = True
-        elif self.failure_check.isChecked():
-            success_only = False
-
-        favorites_only = self.favorites_check.isChecked()
-
-        results, self.total_count = self.history_manager.search_advanced(
-            keyword=keyword,
-            date_from=date_from,
-            date_to=date_to,
-            success_only=success_only,
-            favorites_only=favorites_only,
-            limit=self.ITEMS_PER_PAGE,
-            offset=self.current_offset
-        )
-
-        for item in results:
-            self._add_history_item(item)
-            self._history_items.append(item)
-
-        self.current_offset += len(results)
-
-    def _add_history_item(self, item):
-        """히스토리 항목 추가"""
-        # 즐겨찾기 아이콘
-        fav_icon = "⭐" if item.get('is_favorite', False) else "☆"
-
-        # 타임스탬프
-        timestamp = item.get('timestamp', '')[:16]  # YYYY-MM-DD HH:MM
-
-        # 상태 아이콘
-        status = item.get('status', 'completed')
-        if status == 'pending':
-            status_icon = "⏳"
-        elif status == 'committed':
-            status_icon = "✅"
-        elif status == 'rolled_back':
-            status_icon = "↩️"
-        elif not item.get('success', False):
-            status_icon = "❌"
-        else:
-            status_icon = "✅"
-
-        # 쿼리 미리보기
-        query_preview = item.get('query', '')[:50].replace('\n', ' ')
-        if len(item.get('query', '')) > 50:
-            query_preview += "..."
-
-        # 영향받은 행 수
-        result_count = item.get('result_count', 0)
-        count_str = f"({result_count}행)" if result_count > 0 else ""
-
-        display = f"{fav_icon} {timestamp}  {status_icon} {count_str:>8}  {query_preview}"
-
-        list_item = QListWidgetItem(display)
-        list_item.setData(Qt.ItemDataRole.UserRole, item.get('query', ''))
-        # 항목 ID 저장 (즐겨찾기 토글용)
-        list_item.setData(Qt.ItemDataRole.UserRole + 1, item.get('id') or item.get('timestamp'))
-        self.list_widget.addItem(list_item)
-
-    def _do_search(self):
-        """검색 실행"""
-        self.list_widget.clear()
-        self._history_items.clear()
-        self.current_offset = 0
-        self._is_searching = True
-        self._load_chunk()
-
-    def _reset_search(self):
-        """검색 필터 초기화"""
-        self.keyword_edit.clear()
-        self.date_filter_check.setChecked(False)
-        self.success_check.setChecked(False)
-        self.failure_check.setChecked(False)
-        self.favorites_check.setChecked(False)
-        self.load_history()
-
-    def _toggle_favorite(self):
-        """현재 선택 항목 즐겨찾기 토글"""
-        row = self.list_widget.currentRow()
-        if row < 0:
-            return
-
-        item = self.list_widget.item(row)
-        history_id = item.data(Qt.ItemDataRole.UserRole + 1)
-        if history_id:
-            new_state = self.history_manager.toggle_favorite(history_id)
-
-            # 리스트 항목 텍스트 업데이트
-            text = item.text()
-            if new_state:
-                text = "⭐" + text[1:]
-            else:
-                text = "☆" + text[1:]
-            item.setText(text)
-
-            # 즐겨찾기 수 업데이트
-            self._update_fav_count()
-
-    def _update_fav_count(self):
-        """즐겨찾기 카운트 업데이트"""
-        fav_count = self.history_manager.get_favorite_count()
-        self.fav_count_label.setText(f"⭐ {fav_count}")
-
-    def _show_context_menu(self, pos):
-        """컨텍스트 메뉴 표시"""
-        item = self.list_widget.itemAt(pos)
-        if not item:
-            return
-
-        menu = QMenu(self)
-
-        fav_action = menu.addAction("⭐ 즐겨찾기 토글")
-        fav_action.triggered.connect(self._toggle_favorite)
-
-        copy_action = menu.addAction("📋 쿼리 복사")
-        copy_action.triggered.connect(lambda: self._copy_query(item))
-
-        use_action = menu.addAction("📝 에디터에 붙여넣기")
-        use_action.triggered.connect(self.select_current)
-
-        menu.exec(self.list_widget.mapToGlobal(pos))
-
-    def _copy_query(self, item):
-        """쿼리를 클립보드에 복사"""
-        query = item.data(Qt.ItemDataRole.UserRole)
-        if query:
-            QApplication.clipboard().setText(query)
-
-    def load_more(self):
-        """더 많은 히스토리 로드"""
-        self._load_chunk()
-
-    def update_preview(self, row):
-        """미리보기 업데이트"""
-        if row >= 0:
-            item = self.list_widget.item(row)
-            query = item.data(Qt.ItemDataRole.UserRole)
-            self.preview_text.setPlainText(query)
-        else:
-            self.preview_text.clear()
-
-    def select_query(self, item):
-        """쿼리 선택 (더블클릭)"""
-        query = item.data(Qt.ItemDataRole.UserRole)
-        self.query_selected.emit(query)
-        self.accept()
-
-    def select_current(self):
-        """현재 선택된 쿼리 사용"""
-        item = self.list_widget.currentItem()
-        if item:
-            query = item.data(Qt.ItemDataRole.UserRole)
-            self.query_selected.emit(query)
-            self.accept()
 
 
 # =====================================================================
 # SQL 에디터 탭 (개별 탭 위젯)
 # =====================================================================
-class SQLEditorTab(QWidget):
-    """단일 SQL 에디터 탭"""
-
-    modified_changed = pyqtSignal(bool)  # 수정 상태 변경
-    title_changed = pyqtSignal(str)  # 탭 제목 변경 요청
-
-    def __init__(self, parent=None, tab_index: int = 1):
-        super().__init__(parent)
-        self.file_path = None
-        self.is_modified = False
-        self._tab_index = tab_index
-
-        self._init_ui()
-
-    def _init_ui(self):
-        """UI 초기화"""
-        layout = QVBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        layout.setSpacing(0)
-
-        # 에디터
-        self.editor = ValidatingCodeEditor()
-        self.editor.setPlaceholderText("SELECT * FROM table_name;\n-- Ctrl+Space: 자동완성")
-        self.editor.textChanged.connect(self._on_text_changed)
-        layout.addWidget(self.editor)
-
-        # 검증 상태 라벨
-        self.validation_label = QLabel("")
-        self.validation_label.setStyleSheet("color: #666; font-size: 11px; padding: 2px 4px;")
-        layout.addWidget(self.validation_label)
-
-    def _on_text_changed(self):
-        """텍스트 변경 시"""
-        if not self.is_modified:
-            self.is_modified = True
-            self.modified_changed.emit(True)
-            self.title_changed.emit(self.get_title())
-
-    def get_title(self) -> str:
-        """탭 제목 반환"""
-        if self.file_path:
-            name = os.path.basename(self.file_path)
-        else:
-            name = f"Query {self._tab_index}"
-        return f"{name} *" if self.is_modified else name
-
-    def set_tab_index(self, index: int):
-        """탭 인덱스 설정"""
-        self._tab_index = index
-        self.title_changed.emit(self.get_title())
-
-    def set_content(self, text: str):
-        """내용 설정 (수정 플래그 초기화)"""
-        self._set_large_document_mode_for_text(text)
-        self.editor.blockSignals(True)
-        self.editor.setPlainText(text)
-        self.editor.blockSignals(False)
-        self.is_modified = False
-        self.title_changed.emit(self.get_title())
-
-    def get_content(self) -> str:
-        """내용 반환"""
-        return self.editor.toPlainText()
-
-    def load_file(self, file_path: str) -> bool:
-        """파일 불러오기"""
-        try:
-            with open(file_path, 'r', encoding='utf-8') as f:
-                content = f.read()
-            self._set_large_document_mode_for_text(content)
-            self.editor.blockSignals(True)
-            self.editor.setPlainText(content)
-            self.editor.blockSignals(False)
-            self.file_path = file_path
-            self.is_modified = False
-            self.title_changed.emit(self.get_title())
-            return True
-        except Exception:
-            return False
-
-    def _set_large_document_mode_for_text(self, text: str):
-        byte_size = len(text.encode("utf-8", errors="ignore"))
-        is_large = byte_size >= LARGE_SQL_RENDER_LIMIT_BYTES
-        self.editor.set_large_document_mode(is_large)
-        if is_large:
-            self.validation_label.setText(
-                "대용량 SQL: 구문 하이라이트와 실시간 검증을 비활성화했습니다."
-            )
-        else:
-            self.validation_label.setText("")
-
-    def save_file(self, file_path: str = None) -> tuple:
-        """파일 저장
-
-        Returns:
-            (success, file_path, error_message)
-        """
-        target_path = file_path or self.file_path
-
-        if not target_path:
-            return False, None, "파일 경로가 지정되지 않았습니다."
-
-        try:
-            with open(target_path, 'w', encoding='utf-8') as f:
-                f.write(self.editor.toPlainText())
-            self.file_path = target_path
-            self.is_modified = False
-            self.modified_changed.emit(False)
-            self.title_changed.emit(self.get_title())
-            return True, target_path, None
-        except Exception as e:
-            return False, None, str(e)
-
-    def mark_saved(self):
-        """저장 완료 표시"""
-        self.is_modified = False
-        self.modified_changed.emit(False)
-        self.title_changed.emit(self.get_title())
 
 
 # =====================================================================
